requirements.txt
discord.py
dnspython
python
discord
praw
youtube-dl
aiofiles
Flask
giphy-client
wikipedia
DiscordUtils
PyNaCl
timeago
psutil
base91
opencv-python
Pillow
scikit-image
prsaw
jishaku
lxml
spice-api
colour
sr-apa
snake
dpy-button-utils
discord-ext-games
parsedatetime
aiosqlite
googlesearch-python
discord-better-help